Begin refactoring SymbolTableEntry
