+++ src/main/java/edu/clemson/cs/rsrg/init/CompileEnvironment.java
@@ -21,9 +21,9 @@ import edu.clemson.cs.rsrg.statushandling.exception.MiscErrorException;
 import edu.clemson.cs.rsrg.init.file.ResolveFile;
 import edu.clemson.cs.rsrg.init.flag.FlagManager;
 import edu.clemson.cs.rsrg.misc.Utilities;
-import edu.clemson.cs.rsrg.typeandpopulate.ModuleIdentifier;
 import edu.clemson.cs.rsrg.typeandpopulate.symboltables.ScopeRepository;
 import edu.clemson.cs.rsrg.typeandpopulate.typereasoning.TypeGraph;
+import edu.clemson.cs.rsrg.typeandpopulate.utilities.ModuleIdentifier;
 import java.io.*;
 import java.text.SimpleDateFormat;
 import java.util.*;
