+++ src/main/java/edu/clemson/cs/rsrg/init/pipeline/AbstractPipeline.java
@@ -14,7 +14,7 @@ package edu.clemson.cs.rsrg.init.pipeline;
 
 import edu.clemson.cs.rsrg.init.CompileEnvironment;
 import edu.clemson.cs.rsrg.typeandpopulate.symboltables.MathSymbolTableBuilder;
-import edu.clemson.cs.rsrg.typeandpopulate.ModuleIdentifier;
+import edu.clemson.cs.rsrg.typeandpopulate.utilities.ModuleIdentifier;
 
 /**
  * <p>This is the abstract base class for all pipeline objects
