+++ src/main/java/edu/clemson/cs/rsrg/typeandpopulate/programtypes/PTType.java
@@ -38,7 +38,7 @@ public abstract class PTType {
 
     /**
      * <p>An helper constructor that allow us to store the type graph
-     * of any objects created from a class that inherits from
+     * for any objects created from a class that inherits from
      * {@code PTType}.</p>
      *
      * @param g The current type graph.
