+++ src/main/java/edu/clemson/cs/rsrg/typeandpopulate/utilities/FunctionApplicationFactory.java
@@ -10,7 +10,7 @@
  * This file is subject to the terms and conditions defined in
  * file 'LICENSE.txt', which is part of this source code package.
  */
-package edu.clemson.cs.rsrg.typeandpopulate;
+package edu.clemson.cs.rsrg.typeandpopulate.utilities;
 
 import edu.clemson.cs.rsrg.typeandpopulate.mathtypes.MTFunction;
 import edu.clemson.cs.rsrg.typeandpopulate.mathtypes.MTType;
