+++ src/main/java/edu/clemson/cs/rsrg/typeandpopulate/utilities/ModuleIdentifier.java
@@ -10,7 +10,7 @@
  * This file is subject to the terms and conditions defined in
  * file 'LICENSE.txt', which is part of this source code package.
  */
-package edu.clemson.cs.rsrg.typeandpopulate;
+package edu.clemson.cs.rsrg.typeandpopulate.utilities;
 
 import edu.clemson.cs.rsrg.absyn.declarations.moduledecl.ModuleDec;
 import edu.clemson.cs.rsrg.absyn.items.programitems.UsesItem;
